+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # 2025-03-16
 - #2592 Fixes EVM RPC regression for paymaster enabled rollups
 
+# 2025-03-13
+- #2587 EVM: RPC only: Fixes Fee-cap admission consistency across simulation and submission.
+
 # 2025-03-10
 - #2569 **Breaking change** Updated the `SP1Host` and `SP1Verifier` implementations to support SP1 v6. `ZkvmHost::code_commitment` now returns a Result, which is a breaking change.
 - #2554 **Breaking change** Updated the `SP1Host` and `SP1Verifier` implementations to support SP1 v6. `ZkvmHost::code_commitment` now returns a Result, which is a breaking change.